+++ config/routes.rb
@@ -45,6 +45,9 @@ Rails.application.routes.draw do
 
   namespace :builder do
     resources :surveys do
+      member do
+        get :preview
+      end
       resources :questions do
         collection do
           post :reorder