+++ test/controllers/builder/surveys_controller_test.rb
@@ -1,5 +1,8 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
+# Tests to assure that survey builders can create and update surveys.
 class Builder::SurveysControllerTest < ActionController::TestCase
   setup do
     @builder = users(:builder)
@@ -8,33 +11,33 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     @question = questions(:web_question)
   end
 
-  test "should get index as builder" do
+  test 'should get index as builder' do
     login(@builder)
     get :index
     assert_response :success
   end
 
-  test "should not get index as regular user" do
+  test 'should not get index as regular user' do
     login(@regular_user)
     get :index
     assert_redirected_to root_path
   end
 
-  test "should get new as builder" do
+  test 'should get new as builder' do
     login(@builder)
     get :new
     assert_not_nil assigns(:survey)
     assert_response :success
   end
 
-  test "should not get new as regular user" do
+  test 'should not get new as regular user' do
     login(@regular_user)
     get :new
     assert_nil assigns(:survey)
     assert_redirected_to root_path
   end
 
-  test "should create survey as builder" do
+  test 'should create survey as builder' do
     login(@builder)
     assert_difference('Survey.count') do
       post :create, survey: { name_en: 'My New Survey', slug: 'my-new-survey', status: 'show' }
@@ -46,7 +49,7 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_path(id: assigns(:survey))
   end
 
-  test "should not create survey without name" do
+  test 'should not create survey without name' do
     login(@builder)
     assert_difference('Survey.count', 0) do
       post :create, survey: { name_en: '', slug: 'my-new-survey', status: 'show' }
@@ -58,7 +61,7 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not create survey as regular user" do
+  test 'should not create survey as regular user' do
     login(@regular_user)
     assert_difference('Survey.count', 0) do
       post :create, survey: { name_en: 'My New Survey', slug: 'my-new-survey', status: 'show' }
@@ -67,35 +70,41 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should show survey as builder" do
+  test 'should show survey as builder' do
     login(@builder)
     get :show, id: @survey
     assert_not_nil assigns(:survey)
     assert_response :success
   end
 
-  test "should not show survey as regular user" do
+  test 'should preview survey as builder' do
+    login(@builder)
+    get :preview, id: @survey
+    assert_response :success
+  end
+
+  test 'should not show survey as regular user' do
     login(@regular_user)
     get :show, id: @survey
     assert_nil assigns(:survey)
     assert_redirected_to root_path
   end
 
-  test "should get edit survey as builder" do
+  test 'should get edit survey as builder' do
     login(@builder)
     get :edit, id: @survey
     assert_not_nil assigns(:survey)
     assert_response :success
   end
 
-  test "should not get edit survey as regular user" do
+  test 'should not get edit survey as regular user' do
     login(@regular_user)
     get :edit, id: @survey
     assert_nil assigns(:survey)
     assert_redirected_to root_path
   end
 
-  test "should update survey as builder" do
+  test 'should update survey as builder' do
     login(@builder)
     patch :update, id: @survey, survey: { name_en: 'Updated Survey', slug: 'updated-survey', status: 'hide' }
     assert_not_nil assigns(:survey)
@@ -105,7 +114,7 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_path(id: assigns(:survey))
   end
 
-  test "should not update survey without name" do
+  test 'should not update survey without name' do
     login(@builder)
     patch :update, id: @survey, survey: { name_en: '', slug: 'updated-survey', status: 'hide' }
     assert_not_nil assigns(:survey)
@@ -115,14 +124,14 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not update survey as regular user" do
+  test 'should not update survey as regular user' do
     login(@regular_user)
     patch :update, id: @survey, survey: { name_en: 'Updated Survey', slug: 'updated-survey', status: 'hide' }
     assert_nil assigns(:survey)
     assert_redirected_to root_path
   end
 
-  test "should destroy survey as builder" do
+  test 'should destroy survey as builder' do
     login(@builder)
     assert_difference('Survey.current.count', -1) do
       delete :destroy, id: @survey
@@ -131,7 +140,7 @@ class Builder::SurveysControllerTest < ActionController::TestCase
     assert_redirected_to builder_surveys_path
   end
 
-  test "should not destroy survey as regular user" do
+  test 'should not destroy survey as regular user' do
     login(@regular_user)
     assert_difference('Survey.current.count', 0) do
       delete :destroy, id: @survey